* main: (26 commits)
  Stop consistency scanner while restore is in progress
  Improve logging (#10387)
  Split GLOBAL_TAG_THROTTLING_FOLDING_TIME into several knobs
  Added HoltLinearSmootherImpl::eTrendFoldingTime field
  Use HoltLinearSmoother in GlobalTagThrottler
  Add HoltLinearSmoother
  EaR: Optimize logging from GetEncryptCipherKey (#10326)
  Fix tracing in GlobalTagThrottler::getClientRates
  Fix tracing typo in GlobalTagThrottler
  fixup! joshua_logtool error should be reported as a XML element by test_harness
  Change compiler options order to avoid -gdwarf overriding -g1 (#10377)
  Fix test by disable tenant
  Should repeat when speedUpSimulation is false
  Fix a simulation DR stuck issue
  update links to foundationdb.org to reference GitHub
  add cmake option to include RocksDB Tools with the Rocks DB compile
  SS Audit Storage Throttling (#10322)
  Fix restore range loss
  Fix type of TagCost field in BusiestReadTag traces
  Fix failing /fdbserver/TransactionTagCounter/IgnoreBelowMinRate unit test
  ...

+++ fdbclient/include/fdbclient/FDBTypes.h
@@ -829,7 +829,7 @@ struct RangeResultRef : VectorRef<KeyValueRef> {
 		serializer(ar, ((VectorRef<KeyValueRef>&)*this), more, readThrough, readToBegin, readThroughEnd);
 	}
 
-	int logicalSize() const {
+	int64_t logicalSize() const {
 		return VectorRef<KeyValueRef>::expectedSize() - VectorRef<KeyValueRef>::size() * sizeof(KeyValueRef);
 	}
